requirements.txt
sorl-thumbnail
mptt
redis
Pillow
pgmagick
graphicsmagick
django-ckeditor
unidecode